// src/infra/standard-actions/monitors/monitor-types-enum.ts
export enum MonitorTypeEnum {
  STORAGE_CHECK = 'STORAGE_CHECK',
  FILE_SYSTEM_HEALTH = 'FILE_SYSTEM_HEALTH',
  NETWORK_USAGE = 'NETWORK_USAGE',
  SWAP_USAGE = 'SWAP_USAGE',
  BATTERY_STATUS = 'BATTERY_STATUS',
  MEMORY_CHECK = 'MEMORY_CHECK',
  SYSTEM_LOAD = 'SYSTEM_LOAD',
  CPU_CHECK = 'CPU_CHECK'
}